requirements.txt
pymupdf
pdf2image
pytesseract
tqdm
python-dotenv
langchain
langchain-community
langchain-google-genai
chromadb
sentence-transformers
streamlit
watchdog
